squelch.c optimizations
Added const to the type for some lists that don't change.  This makes
the code that accesses them more efficient.

Noticed that several functions used the same coordinates for displaying
text, and that the variables that held these coordinates were never
modified.  Replaced the many local coordinate variables with a single
pair of defines.